Move primary participant screen
